agents/my_agent/requirements.txt
google-adk==1.1.1
# Required for proprietary models
litellm>=1.63.11
# Required for PostgreSQL database connection
psycopg2-binary>=2.9.9
# Required by google-adk
deprecated>=1.2.14
# Add your additional dependencies below:
# requests>=2.31.0
# beautifulsoup4>=4.12.0
# numpy>=1.24.0